unia pól, wydzielenie wyspecjalizowanych fragmentów pól, usunięcie FieldFragmentu
